% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackjack.R
\name{simulate_blackjack}
\alias{simulate_blackjack}
\title{Simulate Blackjack Games}
\usage{
simulate_blackjack(n_rounds = 1000)
}
\arguments{
\item{n_rounds}{Number of rounds to simulate}
}
\value{
A tibble summarizing win/loss/draw counts
}
\description{
Simulate Blackjack Games
}
\examples{
simulate_blackjack(1000)
}
