% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{simulate_dice}
\alias{simulate_dice}
\title{Simulate Dice Rolls}
\usage{
simulate_dice(n = 1000, sides = 6)
}
\arguments{
\item{n}{Number of rolls}

\item{sides}{Number of sides on the dice (default is 6)}
}
\value{
A tibble with frequency counts of outcomes
}
\description{
Simulate Dice Rolls
}
\examples{
simulate_dice(1000)
}
