% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall.R
\name{simulate_monty_hall}
\alias{simulate_monty_hall}
\title{Simulate the Monty Hall Problem}
\usage{
simulate_monty_hall(n = 1000, switch = TRUE)
}
\arguments{
\item{n}{Number of simulations}

\item{switch}{Logical, whether the player switches doors}
}
\value{
A tibble with win/loss counts and win rate
}
\description{
Simulate the Monty Hall Problem
}
\examples{
simulate_monty_hall(1000, switch = TRUE)
}
