% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock_paper_scissor.R
\name{simulate_rps}
\alias{simulate_rps}
\title{Simulate Rock-Paper-Scissors}
\usage{
simulate_rps(n = 1000)
}
\arguments{
\item{n}{Number of games to simulate}
}
\value{
A tibble summarizing win/loss/draw counts for player 1
}
\description{
Simulate Rock-Paper-Scissors
}
\examples{
simulate_rps(1000)
}
